fix:(mediacontrol) enabling and disabling an outsider scrollview when the user swipe the slider of the player.
This function was necessary because our apps were using the player inside
a scrollview, so a conflict was happening between the swipe event in the
slider and the scrollview.

The comment was necessary because it is a very uncommon situation, we intend to change the MediaControl slide bar
for a native UIControl.