refactor: Reorganize files to be more maintanable Current structure of folders and files: Clappr > ... Clappr_tvOS > ...
New structure of folders and files:
Clappr_iOS > ...
Clappr_tvOS > ...
Clappr > ...

Advantages:

In the current structure, if a new class is added only in the iOS target, it's mandatory to add on exclude_files. In the new structure, It's not necessary to add the reference on the podspec.
The files that only exist in iOS will be more visible